remove model open settings in newtab
