+++ css/newtab.css
@@ -1,5 +1,5 @@
 * {
-  font-family: 'Salsa', 'Roboto', sans-serif;
+  font-family: "Salsa", "Roboto", sans-serif;
 }
 
 body {
@@ -19,34 +19,16 @@ body {
   color: #333;
 }
 
-.setting {
-  display: flex;
-  align-items: center;
-}
-input {
-  padding-left: 4px;
-  border: solid 1px;
-  border-radius: 7px;
-  margin: 2px 0px;
-}
-
-select {
-  padding-left: 4px;
-  border: solid 1px;
-  border-radius: 7px;
-  margin: 2px 0px;
-}
-
 .background-container {
   position: fixed;
   top: 0;
   left: 0;
-  width: 100vw; /* Full viewport width */
-  height: 100vh; /* Full viewport height */
-  background-image: url("/img/noise/noise-dark.png"); /* Replace with the path to your 50px * 50px image */
-  background-size: 220px; /* Size of your small image */
-  background-repeat: repeat; /* Repeat the image both horizontally and vertically */
-  z-index: -1; /* Place the background behind other elements */
+  width: 100vw;
+  height: 100vh;
+  background-image: url("/img/noise/noise-dark.png");
+  background-size: 220px;
+  background-repeat: repeat;
+  z-index: -1;
 }
 
 #quick-settings {
@@ -66,66 +48,8 @@ select {
   opacity: 1;
 }
 
-.show-modal {
-  width: 100px;
+.show-settings {
+  width: 24px;
   height: auto;
   margin-bottom: 12px;
-}
-
-.close-modal {
-  position: absolute;
-  top: 1.2rem;
-  right: 2rem;
-  font-size: 2.5rem;
-  color: #333;
-  cursor: pointer;
-  border: none;
-  background: none;
-}
-
-.hidden {
-  display: none;
-}
-
-.modal {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 90%;
-  height: 90%;
-  background-color: white;
-  padding: 1rem 2rem;
-  border-radius: 5px;
-  box-shadow: 0 3rem 5rem rgba(0, 0, 0, 0.3);
-  z-index: 10;
-}
-
-.overlay {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.6);
-  backdrop-filter: blur(3px);
-  z-index: 5;
-}
-
-.modal-content {
-  margin-top: 12px;
-}
-
-.tab-selection {
-  margin: 0px 2px;
-  padding: 1px 6px;
-  border-radius: 2px;
-}
-
-.tab-selection:hover {
-  background-color: aquamarine;
-}
-
-.tab-selection-selected {
-  background-color: aquamarine;
 }
