--- css/popup.css
@@ -1,81 +1,0 @@
-* {
-  font-family: 'Salsa', 'Roboto', system-ui, -apple-system, BlinkMacSystemFont,
-    'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue',
-    sans-serif;
-  margin: 1px;
-}
-
-html {
-  height: 500px;
-  width: 300px;
-  margin: 0;
-}
-
-body {
-  margin: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  /* align-items: center; */
-  height: fit-content;
-  width: 100%;
-  background: linear-gradient(135deg, #88cbd9, #bfecab, #ffffff);
-  padding: 3%;
-
-}
-
-hr {
-  color: #333;
-  border: 0.5px solid;
-  width: 92%;
-}
-
-.main-heading {
-  font-size: medium;
-}
-
-.clock {
-  width: fit-content;
-  font-size: large;
-  font-family: 'Arial', sans-serif;
-  color: #333;
-  padding: 10%;
-  border: 0.2px solid #333;
-}
-
-#options-container {
-  display: flex;
-  flex-direction: column;
-}
-
-#font-selection {
-  display: flex;
-  flex-direction: column;
-}
-
-#googleFontInput {
-  width: fit-content;
-}
-
-#applyGoogleFont {
-  width: fit-content;
-}
-
-#buymeacoffee {
-  width: 150px;
-  height: auto;
-}
-
-#goToFullSettings {
-  width: 93%;
-  color: red;
-  border: solid 1px red;
-  border-radius: 7px;
-  background-color: blanchedalmond;
-}
-
-#goToFullSettings:hover {
-  color: black;
-  border-color: black;
-  background: #ddd;
-}
