+++ settings.html
@@ -1,0 +1,95 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./css/settings.css" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Salsa&family=Long+Cang&family=Roboto&display=swap"
+      rel="stylesheet"
+    />
+    <title>Settings ✨</title>
+    <script defer src="./js/settings.js" type="module"></script>
+  </head>
+  <body>
+    <h1 id="main-heading">General Settings ✨</h1>
+    <hr style="border: solid 1px" />
+
+    <div id="section-headings" style="display: flex">
+      <div class="tab-selection tab-selection-selected" id="link1-div">
+        <a href="#" id="link1">General</a>
+      </div>
+      <div class="tab-selection" id="link2-div">
+        <a href="#" id="link2">Extensions</a>
+      </div>
+      <div class="tab-selection" id="link3-div">
+        <a href="#" id="link3">Attributes</a>
+      </div>
+    </div>
+
+    <div id="content1">
+      <div class="setting">
+        <input type="checkbox" id="auto-save-settings" />
+        <label for="auto-save-settings">Auto Save Settings</label>
+      </div>
+
+      <div>
+        <h2>Clock Settings</h2>
+        <div class="setting">
+          <input type="checkbox" id="showMilliseconds" />
+          <label for="showMilliseconds">Show Seconds </label>
+        </div>
+
+        <div class="setting">
+          <input type="checkbox" id="use12HourFormat" />
+          <label for="use12HourFormat">Use 12-Hour Clock</label>
+        </div>
+      </div>
+
+      <div>
+        <h2>Tab Settings</h2>
+        <div class="setting">
+          <label for="newTabName">Tab Title:</label>
+          <input type="text" id="newTabName" />
+        </div>
+      </div>
+
+      <div>
+        <h2>Font Settings</h2>
+        <div class="setting">
+          <label for="fontSize">Font Size:</label>
+          <input type="number" id="fontSize" />
+        </div>
+        <div class="setting">
+          <label for="fontFamily">Font Family:</label>
+          <select id="fontFamily">
+            <option value="Roboto">Roboto</option>
+            <option value="Salsa">Salsa</option>
+            <option value="Long Cang">Long Cang</option>
+            <option value="Sixtyfour">Sixtyfour</option>
+          </select>
+        </div>
+      </div>
+
+      <div>
+        <h2>Refresh Rate</h2>
+        <div class="setting">
+          <label for="refreshRate">Refresh Rate:</label>
+          <input type="number" id="refreshRate" />
+        </div>
+      </div>
+    </div>
+
+    <div id="content2">
+      <div class="setting">
+        <input type="checkbox" id="sync-profiles" />
+        <label for="sync-profiles">Sync across Chrome Profiles</label>
+      </div>
+
+      <div class="setting">
+        <input type="checkbox" id="realtime-sync" />
+        <label for="realtime-sync">Sync settings in realtime</label>
+      </div>
+    </div>
+  </body>
+</html>
